fix: rename mismatched and duplicate filenames
