+++ lib/card_templates/bulk/bulk_generation_screen.dart
@@ -5,10 +5,10 @@ import 'dart:ui' as ui;
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 import 'package:image/image.dart' as img;
-import 'package:magicepaperapp/card_templates/bulk/bulk_result.dart';
+import 'package:magicepaperapp/card_templates/bulk/generated_badge.dart';
 import 'package:magicepaperapp/card_templates/bulk/bulk_results_screen.dart';
 import 'package:magicepaperapp/card_templates/bulk/bulk_template.dart';
-import 'package:magicepaperapp/card_templates/bulk/photo_source.dart';
+import 'package:magicepaperapp/card_templates/bulk/photo_resolver.dart';
 import 'package:magicepaperapp/constants/color_constants.dart';
 import 'package:magicepaperapp/constants/dimens.dart';
 import 'package:magicepaperapp/l10n/app_localizations.dart';
