+++ lib/card_templates/bulk/generated_badge.dart
