+++ lib/card_templates/bulk/photo_resolver.dart
