+++ lib/card_templates/contact_card_form.dart
@@ -3,7 +3,7 @@ import 'dart:typed_data';
 import 'package:flutter/material.dart';
 import 'package:magicepaperapp/card_templates/utils/image_picker_util.dart';
 import 'package:magicepaperapp/card_templates/card_template_result.dart';
-import 'package:magicepaperapp/card_templates/contact_card_card_widget.dart';
+import 'package:magicepaperapp/card_templates/contact_card_widget.dart';
 import 'package:magicepaperapp/card_templates/contact_card_model.dart';
 import 'package:magicepaperapp/card_templates/template_layer_builders.dart';
 import 'package:magicepaperapp/constants/color_constants.dart';
